fix: [UpdateAffiliateInformation] Extract UFPForm in prep for generation of defaultValues
